+++ nilearn/surface/__init__.py
@@ -1,0 +1,7 @@
+"""
+Functions for surface manipulation.
+"""
+
+from .surface import vol_to_surf, load_surf_data, load_surf_mesh
+
+__all__ = ['vol_to_surf', 'load_surf_data', 'load_surf_mesh']